+++ app/src/main/java/net/dom53/inkita/core/network/KavitaApiFactory.kt
@@ -2,10 +2,10 @@ package net.dom53.inkita.core.network
 
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import net.dom53.inkita.core.network.NetworkLoggingInterceptor
 import net.dom53.inkita.data.api.KavitaApi
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.converter.scalars.ScalarsConverterFactory
@@ -30,15 +30,10 @@ object KavitaApiFactory {
     }
 
     fun createUnauthenticated(baseUrl: String): KavitaApi {
-        val logging =
-            HttpLoggingInterceptor().apply {
-                level = HttpLoggingInterceptor.Level.BODY
-            }
-
         val client =
             OkHttpClient
                 .Builder()
-                .addInterceptor(logging)
+                .addInterceptor(NetworkLoggingInterceptor)
                 .build()
 
         val retrofit = baseRetrofitBuilder(baseUrl, client).build()
@@ -60,16 +55,11 @@ object KavitaApiFactory {
                 chain.proceed(newRequest)
             }
 
-        val logging =
-            HttpLoggingInterceptor().apply {
-                level = HttpLoggingInterceptor.Level.BODY
-            }
-
         val client =
             OkHttpClient
                 .Builder()
                 .addInterceptor(apiKeyInterceptor)
-                .addInterceptor(logging)
+                .addInterceptor(NetworkLoggingInterceptor)
                 .build()
 
         val retrofit = baseRetrofitBuilder(baseUrl, client).build()